requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=1.5.0
tqdm>=4.65.0
umap-learn>=0.5.4
